src/cloud-functions/air_pollution/requirements.txt
blinker==1.9.0 ; python_version >= "3.11" and python_version < "4.0"
cachetools==5.5.2 ; python_version >= "3.11" and python_version < "4.0"
certifi==2025.1.31 ; python_version >= "3.11" and python_version < "4.0"
charset-normalizer==3.4.1 ; python_version >= "3.11" and python_version < "4.0"
click==8.1.8 ; python_version >= "3.11" and python_version < "4.0"
cloudevents==1.11.0 ; python_version >= "3.11" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.11" and python_version < "4.0" and platform_system == "Windows"
deprecated==1.2.18 ; python_version >= "3.11" and python_version < "4.0"
deprecation==2.1.0 ; python_version >= "3.11" and python_version < "4.0"
flask==3.1.0 ; python_version >= "3.11" and python_version < "4.0"
functions-framework==3.8.2 ; python_version >= "3.11" and python_version < "4.0"
geographiclib==2.0 ; python_version >= "3.11" and python_version < "4.0"
geopy==2.4.1 ; python_version >= "3.11" and python_version < "4.0"
google-api-core==2.24.2 ; python_version >= "3.11" and python_version < "4.0"
google-auth==2.38.0 ; python_version >= "3.11" and python_version < "4.0"
google-cloud-pubsub==2.29.0 ; python_version >= "3.11" and python_version < "4.0"
googleapis-common-protos==1.69.2 ; python_version >= "3.11" and python_version < "4.0"
grpc-google-iam-v1==0.14.2 ; python_version >= "3.11" and python_version < "4.0"
grpcio-status==1.71.0 ; python_version >= "3.11" and python_version < "4.0"
grpcio==1.71.0 ; python_version >= "3.11" and python_version < "4.0"
gunicorn==23.0.0 ; python_version >= "3.11" and python_version < "4.0" and platform_system != "Windows"
idna==3.10 ; python_version >= "3.11" and python_version < "4.0"
importlib-metadata==8.6.1 ; python_version >= "3.11" and python_version < "4.0"
itsdangerous==2.2.0 ; python_version >= "3.11" and python_version < "4.0"
jinja2==3.1.6 ; python_version >= "3.11" and python_version < "4.0"
markupsafe==3.0.2 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-api==1.31.1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-sdk==1.31.1 ; python_version >= "3.11" and python_version < "4.0"
opentelemetry-semantic-conventions==0.52b1 ; python_version >= "3.11" and python_version < "4.0"
packaging==24.2 ; python_version >= "3.11" and python_version < "4.0"
proto-plus==1.26.1 ; python_version >= "3.11" and python_version < "4.0"
protobuf==5.29.4 ; python_version >= "3.11" and python_version < "4.0"
pyasn1-modules==0.4.1 ; python_version >= "3.11" and python_version < "4.0"
pyasn1==0.6.1 ; python_version >= "3.11" and python_version < "4.0"
requests==2.32.3 ; python_version >= "3.11" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.11" and python_version < "4.0"
typing-extensions==4.12.2 ; python_version >= "3.11" and python_version < "4.0"
urllib3==2.3.0 ; python_version >= "3.11" and python_version < "4.0"
watchdog==6.0.0 ; python_version >= "3.11" and python_version < "4.0"
werkzeug==3.1.3 ; python_version >= "3.11" and python_version < "4.0"
wrapt==1.17.2 ; python_version >= "3.11" and python_version < "4.0"
zipp==3.21.0 ; python_version >= "3.11" and python_version < "4.0"
